#172 - Create a CityLabelScene that handles the complexity of the label.
This is towards a goal of having relatively self-contained units.  We might have a CityFireworksScene as well some day.
